Revert "refactor: ProfitCalculator 패키지 변경"
This reverts commit 99716a81961e4150463fc58f8abd0f6a2d7c2356.
