+++ src/main/java/blackjack/domain/result/ResultTable.java
@@ -1,8 +1,5 @@
 package blackjack.domain.result;
 
-import blackjack.domain.bet.BettingTable;
-import blackjack.domain.bet.Money;
-
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -25,39 +22,5 @@ public class ResultTable {
 
     public List<String> getGameEndedPlayerNames() {
         return List.copyOf(table.keySet());
-    }
-
-    public static class ProfitCalculator {
-
-        private final BettingTable bettingTable;
-        private final ResultTable resultTable;
-
-        public ProfitCalculator() {
-            this.bettingTable = new BettingTable();
-            this.resultTable = new ResultTable();
-        }
-
-        public void bet(final String name, final int betAmount) {
-            bettingTable.bet(name, new Money(betAmount));
-        }
-
-        public void putGameResult(final String name, final GameResult result) {
-            resultTable.put(name, result);
-        }
-
-        public int calculateProfit(final String name) {
-            final GameResult gameResult = resultTable.get(name);
-            return gameResult.calculateProfit(
-                    bettingTable.get(name).getAmount()
-            );
-        }
-
-        public int calculateDealerProfit() {
-            final List<String> playerNames = resultTable.getGameEndedPlayerNames();
-            final int playerProfitSum = playerNames.stream()
-                    .mapToInt(this::calculateProfit)
-                    .sum();
-            return -1 * playerProfitSum;
-        }
     }
 }
